
\name{data}
\alias{data}
\alias{logR}
\alias{logOR}
\alias{chrLoc}

%\docType{package}
\title{
  Data for examples
}
\description{
 Data for examples.
}
\details{ The object contains \code{logR}, \code{logOR}, and \code{chrLoc}
 for \code{\link{subhmm}}.
}
\examples{

 data(data, package="subHMM")

 # Display some of the data
 logR[1:5]
 logOR[1:5]
 chrLoc[1:5, ]
 
}
\keyword{data}
\seealso{ \code{\link{subhmm}} }



